requirements.txt
# Transcription
assemblyai>=0.30.0

# Database
psycopg2-binary>=2.9.9

# Voyage AI embeddings
requests>=2.31.0

# YAML parsing for frontmatter
pyyaml>=6.0
